lib: use uint64_t for page offset in mmap helpers
The page offset is computed as a uint64_t address modulo a uint64_t
page_size, but was stored in a uint32_t, silently truncating the
result. Use uint64_t for offset in pqos_mmap_read and pqos_mmap_write
to avoid the implicit truncation.

+++ lib/common.c
@@ -305,7 +305,7 @@ pqos_file_contains(const char *fname, const char *str, int *found)
 uint8_t *
 pqos_mmap_read(uint64_t address, const uint64_t size)
 {
-        uint32_t offset;
+        uint64_t offset;
         uint64_t page_size;
         uint8_t *mem;
         int fd;
@@ -336,7 +336,7 @@ pqos_mmap_read(uint64_t address, const uint64_t size)
 uint8_t *
 pqos_mmap_write(uint64_t address, const uint64_t size)
 {
-        uint32_t offset;
+        uint64_t offset;
         uint64_t page_size;
         uint8_t *mem;
         int fd;
